Prepare 1.1.2 release
- [BUG] Add missing function entry termination for TLData
- [CLEANUP] Don't return garbage on hard class initialization errors, don't throw from create_object 
handlers

        



git-svn-id: http://svn.php.net/repository/pecl/krb5/trunk@342263 c90b9560-bf6c-de11-be94-00142212c4b1

+++ testing.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-PHP_VERSIONS="5.6 7.0"
+PHP_VERSIONS="5.6 7.1"
 VERSIONS="1.8 1.9 1.10 1.11 1.12"
 
 for PHP_VERSION in $PHP_VERSIONS
@@ -14,9 +14,24 @@ do
 		test -f Makefile && make distclean &> /dev/null
 		phpize --clean &> /dev/null &&
 		$PHP_PATH/bin/phpize
-		( aclocal; autoconf; autoheader; automake
+		rm -f .libs/krb5.so
+
+		if [ "$PHP_VERSION" == "7.1" ]
+		then
+			REGEN=false
+		else
+			REGEN=true
+		fi
+		( $REGEN && aclocal; $REGEN && autoconf; $REGEN && autoheader; $REGEN && automake
 		./configure --with-php-config=$PHP_PATH/bin/php-config --with-krb5config=$KRB5DIR/bin/krb5-config --with-krb5kadm &&
-		make ) 2>&1 > test.${PHP_VERSION}-${VERSION}.build.out &&
+		make clean && make ) 2>&1 > test.${PHP_VERSION}-${VERSION}.build.out
+
+		if [ ! -f .libs/krb5.so ]
+		then
+			echo "Failed KRB5 $VERSION PHP $PHP_VERSION"
+			exit
+		fi
+
 		cp testing.config.php tests/config.php &&
 		make test NO_INTERACTION=yes &> test.${PHP_VERSION}-${VERSION}.out
 		cat test.${PHP_VERSION}-${VERSION}.out
